% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_rr_options.R
\name{list_rr_options}
\alias{list_rr_options}
\title{Provide "stock" and "management_unit_name" options for a run reconstruction dataframe}
\usage{
list_rr_options(data, stock = NULL)
}
\arguments{
\item{data}{Dataframe generated by \code{read_rr_trs()}}

\item{stock}{Optional, stock}
}
\value{
character vector
}
\description{
Once a run reconstruction dataframe is created, it's nice to see the options for \code{stock} and
\code{management_unit_name} identifiers for making plots. This function acts as a wrapper for
\code{unique(data$stock)} if no stock name is provided; if stock name is provided,
instead is a wrapper for filtering to that stock and then \verb{unique(data$management_unit_name}.
When presenting management unit names, removes astrixes; this is consistent with \code{plot_rr_mu()} behavior.
}
