% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrViewer-package.R
\docType{package}
\name{rrViewer-package}
\alias{rrViewer}
\alias{rrViewer-package}
\title{rrViewer: Plotting tools for evaluating Coho run reconstructions}
\description{
One of the yearly tasks of the FRAM team is to complete the Coho Run Reconstruction. This involves careful evaluation of myriad values in excel files to confirm that they are within reasonable expectations. This package assists in this process by plotting the values of the current run reconstruction across equivalent values in past years.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cbedwards-dfw/rrViewer}
  \item \url{https://cbedwards-dfw.github.io/rrViewer/}
  \item Report bugs at \url{https://github.com/cbedwards-dfw/rrViewer/issues}
}

}
\author{
\strong{Maintainer}: Collin Edwards \email{Collin.Edwards@dfw.wa.gov} (\href{https://orcid.org/0000-0002-4937-5159}{ORCID})

}
\keyword{internal}
